requirements.txt
streamlit==1.12.0
pandas
openpyxl
python-docx
python-pptx
groq
python-dotenv
sentence-transformers
numpy
scikit-learn
langchain-core
langchain_groq
langchain_huggingface
ipykernel